feat(CreateFab): add create fab component and remove from layout
